+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 Warning: Before deploying in production a bot that can generate heavy traffic, please contact ALE.
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
+## [1.6x.x] - 2020-01-xx
+-   Fix typo error in Gruntfile.js file for the generated documentation of the invitations service. And fix documentation.
+-   Move the methods BubblesService::resizeImage and BubblesService::getBinaryData to module Utils.ts
+-   Add method FileStorageService::uploadFileToStorage to Send a file in user storage. 
+
 ## [1.66.1] - 2020-01-29
 -   Fix when the SDK is already stopped when stop method is called, then return a succeed. (CRRAINB-10270: CPaaS Node SDK - Chief bot demo wasn't unable to restart after connection issue)
 -   Add BubblesService::getUsersFromBubble to get the actives users of a bubble.